base patch for user defined chains support

+++ kernel/linux/include/linux/netfilter_bridge/ebt_redirect.h
@@ -3,8 +3,8 @@
 
 struct ebt_redirect_info
 {
-	// EBT_ACCEPT, EBT_DROP or EBT_CONTINUE
-	__u8 target;
+	// EBT_ACCEPT, EBT_DROP or EBT_CONTINUE or EBT_RETURN
+	int target;
 };
 #define EBT_REDIRECT_TARGET "redirect"
